+++ atlasdb-tests-shared/src/main/java/com/palantir/atlasdb/transaction/impl/AbstractSerializableTransactionTest.java
@@ -66,7 +66,7 @@ public abstract class AbstractSerializableTransactionTest extends AbstractTransa
 
     @Override
     protected TransactionManager getManager() {
-        return SerializableTransactionManager.createForTest(
+        return SerializableTransactionManagerImpl.createForTest(
                 keyValueService,
                 timestampService,
                 lockClient,